move linear-unary UT to test_torchinductor
